Feat: Adds last clicked timestamp to bookmarks
Adds a "last_clicked" timestamp column to the bookmarks table.
This allows tracking the last time a user clicked on a bookmark.

The "goto" endpoint is updated to set the last clicked field
whenever a user navigates to a bookmark.

Also updates Drizzle schema and database migration to reflect the change.